Build su-exec into image (#402)
* Build su-exec into image

su-exec is typically used to drop root privileges in an entrypoint.  We
use it as a replacement for sudo, as sudo setup in a container is a bit
tricky. The su-exec tool is necessary in order to chown /home to the UID
of the user when not running as root.  In this case, chown requires EUID=0.
To grant that, we use su-exec, which is effectively a passwordless sudo sans
a bunch of configuration.

* Address reviewer comments

Move su-exec into the build_tools_contexst
Additionally sort the environments for the build_tools_context