Refactor istio-install setup (#916)
* Refactor istio-install setup

This brings a number of changes to simplify and make the istio install
more flexible. This removes the legacy helm setup, which is no longer
used in master (for old installs, use other branches of tools).
Additionally, it allows installing istio from any source. This is useful
for testing private builds or distributions. This also allows simple
options to configure istio with different profiles or --set flags.

Finally, the prometheus operator setup is cleaned up and merged with the
other "extras". This fixes https://github.com/istio/istio/issues/23262
in the process.

The changes here are documented in the README, please start there.

* fix lint

* minor issues

* fix dry run

* fix lint

* Fix setup